carla_rentals/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.21.45
botocore==1.24.45
Django==4.0.4
django-environ==0.8.1
django-rest-framework==0.1.0
django-storages==1.12.3
djangorestframework==3.13.1
env==0.1.0
environ==1.0
jmespath==1.0.0
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
